Conflicts:
	scilab/modules/gui/help/en_US/uicontrol.xml

+++ scilab/modules/xcos/src/java/org/scilab/modules/xcos/block/io/ContextUpdate.java
@@ -12,17 +12,25 @@
 
 package org.scilab.modules.xcos.block.io;
 
+import java.beans.PropertyChangeEvent;
+import java.beans.PropertyChangeListener;
 import java.io.File;
 import java.io.IOException;
+import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.EnumMap;
 import java.util.List;
 import java.util.Map;
+import java.util.regex.Pattern;
 
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
 import org.scilab.modules.action_binding.highlevel.ScilabInterpreterManagement;
 import org.scilab.modules.action_binding.highlevel.ScilabInterpreterManagement.InterpreterException;
-import org.scilab.modules.types.scilabTypes.ScilabDouble;
-import org.scilab.modules.types.scilabTypes.ScilabList;
+import org.scilab.modules.types.ScilabDouble;
+import org.scilab.modules.types.ScilabList;
+import org.scilab.modules.types.ScilabString;
+import org.scilab.modules.types.ScilabType;
 import org.scilab.modules.xcos.block.BasicBlock;
 import org.scilab.modules.xcos.block.SuperBlock;
 import org.scilab.modules.xcos.graph.SuperBlockDiagram;
@@ -35,24 +43,130 @@ import org.scilab.modules.xcos.port.input.ImplicitInputPort;
 import org.scilab.modules.xcos.port.output.ExplicitOutputPort;
 import org.scilab.modules.xcos.port.output.ImplicitOutputPort;
 import org.scilab.modules.xcos.utils.FileUtils;
+import org.scilab.modules.xcos.utils.XcosEvent;
 
 import com.mxgraph.model.mxGraphModel;
 import com.mxgraph.model.mxICell;
 import com.mxgraph.model.mxIGraphModel;
+import com.mxgraph.util.mxEventObject;
 
 /**
  * Common class for the SuperBlock I/O blocks (represent ports)
  */
 public abstract class ContextUpdate extends BasicBlock {
 
+	private static final Log LOG_LOCAL = LogFactory.getLog(ContextUpdate.class);
 	private static final long serialVersionUID = 6076826729067963560L;
 
+	/**
+	 * Implement a listener which update the value and refresh the view when the
+	 * index of the port change.
+	 */
+	private static final class IndexChangeAdapter implements PropertyChangeListener, Serializable {
+		private static IndexChangeAdapter instance;
+		
+		/**
+		 * Default constructor.
+		 */
+		private IndexChangeAdapter() { }
+		
+		/**
+		 * @return the instance
+		 */
+		public static synchronized IndexChangeAdapter getInstance() {
+			if (instance == null) {
+				instance = new IndexChangeAdapter();
+			}
+			return instance;
+		}
+		
+		/**
+		 * Update the value and refresh the graph view.
+		 * 
+		 * @param evt the event
+		 * @see java.beans.PropertyChangeListener#propertyChange(java.beans.PropertyChangeEvent)
+		 */
+		@Override
+		public void propertyChange(PropertyChangeEvent evt) {
+			ScilabType data = (ScilabType) evt.getNewValue();
+			ContextUpdate ioBlock = (ContextUpdate) evt.getSource();
+			
+			if (!data.isEmpty()) {
+				int newIndex = (int) ((ScilabDouble) data).getRealPart()[0][0];
+				
+				int oldIndex;
+				if (evt.getOldValue() instanceof ScilabDouble
+						&& !((ScilabDouble) evt.getOldValue()).isEmpty()) {
+					oldIndex = (int) ((ScilabDouble) evt.getOldValue()).getRealPart()[0][0];
+				} else {
+					oldIndex = -1;
+				}
+				
+				ioBlock.setValue(newIndex);
+				
+				if (ioBlock.getParentDiagram() != null) {
+					ioBlock.getParentDiagram().fireEvent(
+							new mxEventObject(XcosEvent.IO_PORT_VALUE_UPDATED,
+									"block", ioBlock, "oldIndex", oldIndex,
+									"newIndex", newIndex));
+				}
+			}
+		}
+		
+	}
+	
+	/**
+	 * Implement a listener to update the {@link ContextUpdate#isContextDependent} flag.
+	 */
+	private static final class ExprsChangeAdapter implements PropertyChangeListener, Serializable {
+		private static final Pattern INTEGER_PATTERN = Pattern.compile("\\d+");
+
+		private static ExprsChangeAdapter instance;
+		
+		/**
+		 * Default constructor
+		 */
+		public ExprsChangeAdapter() { }
+		
+		/**
+		 * @return the shared instance
+		 */
+		public static ExprsChangeAdapter getInstance() {
+			if (instance == null) {
+				instance = new ExprsChangeAdapter();
+			}
+			return instance;
+		}
+		
+		/**
+		 * isContextDependant field
+		 * 
+		 * @param evt the event
+		 * @see java.beans.PropertyChangeListener#propertyChange(java.beans.PropertyChangeEvent)
+		 */
+		@Override
+		public void propertyChange(final PropertyChangeEvent evt) {
+			final ScilabType data = (ScilabType) evt.getNewValue();
+			final ContextUpdate ioBlock = (ContextUpdate) evt.getSource();
+			
+			if (!data.isEmpty()) {
+				final String newIndex = ((ScilabString) data).getData()[0][0];
+				
+				if (!INTEGER_PATTERN.matcher(newIndex).matches()) {
+					ioBlock.isContextDependent = true;
+				} else {
+					ioBlock.isContextDependent = false;
+				}
+			}
+		}
+	}
+	
 	/**
 	 * This enum represent all the subclasses of ContextUpdate .
 	 * 
 	 * It is used to easily loop over a BasicBlock I/O blocks
 	 */
-	public enum IOBlocks {
+	public static enum IOBlocks {
 		/** Map a control port to an event input block */
 		EventInBlock(EventInBlock.class, ControlPort.class),
 		/** Map a command port to an event output block */
@@ -141,8 +255,8 @@ public abstract class ContextUpdate extends BasicBlock {
 			final Object[] children = mxGraphModel.getChildCells(
 					defaultModel, defaultParent, true, false);
 			
-			for (int i = 0; i < children.length; i++) {
-				final mxICell child = (mxICell) children[i];
+			for (Object element : children) {
+				final mxICell child = (mxICell) element;
 
 				/* if compatible add it to the list */
 				for (IOBlocks b : IOBlocks.values()) {
@@ -169,21 +283,19 @@ public abstract class ContextUpdate extends BasicBlock {
 			return port;
 		}
 	}
-
+	
+	private transient boolean isContextDependent;
+	
 	/**
-	 * Constructor
+	 * Constructor.
 	 */
 	public ContextUpdate() {
 		super();
-	}
-
-	/**
-	 * @param label
-	 *            block labek
-	 */
-	protected ContextUpdate(String label) {
-		this();
-		setValue(label);
+		
+		getParametersPCS().addPropertyChangeListener(INTEGER_PARAMETERS,
+				IndexChangeAdapter.getInstance());
+		getParametersPCS().addPropertyChangeListener(EXPRS,
+				ExprsChangeAdapter.getInstance());
 	}
 
 	/**
@@ -208,9 +320,16 @@ public abstract class ContextUpdate extends BasicBlock {
 			return;
 		}
 
-		final File tempOutput;
-		final File tempInput;
-		final File tempContext;
+		// do not evaluate context is the block is not context dependent.
+		if (!isContextDependent) {
+			return;
+		}
+		
+		LOG_LOCAL.trace("Update the I/O value from the context");
+		
+		File tempOutput;
+		File tempInput;
+		File tempContext;
 		try {
 			tempInput = FileUtils.createTempFile();
 			tempInput.deleteOnExit();
@@ -233,7 +352,7 @@ public abstract class ContextUpdate extends BasicBlock {
 			updateBlockSettings(modifiedBlock);
 
 		} catch (IOException e) {
-			e.printStackTrace();
+			LOG_LOCAL.error(e);
 		}
 	}
 }
